refactor(orchestrator): share readiness probe, backend-neutral artifact error
Move wait_ready/http_probe from the docker backend to the module root so
upcoming backends reuse them, and rename ImageMissing to ArtifactMissing
since non-container backends miss binaries, not images.
